+++ info.rkt
@@ -11,10 +11,12 @@
 (define drracket-tool-icons '(#f #f #f #f))
 
 (define collection "sauron")
-(define deps '("base" "gui-lib" "drracket-plugin-lib" "from-template"
-               "file-watchers"))
+(define deps '("base" "gui-lib" "net-lib"
+                      "drracket-plugin-lib"
+                      "from-template"
+                      "file-watchers"))
 (define build-deps '("scribble-lib" "racket-doc" "rackunit-lib"
-                     "gui-doc"))
+                                    "gui-doc"))
 (define scribblings '(("scribblings/sauron.scrbl" (multi-page) (tool))))
 (define pkg-desc "A Racket IDE")
 (define version "1.0.4")